Transform motor speed from int to float

+++ src/tests/DFR0592/dfr0592.h
@@ -131,7 +131,7 @@ int encoder_get_speed(const struct dfr_board *board, int motor, int *speed);
  *
  *
  */
-int motor_set_speed(const struct dfr_board *board, int motor, int speed);
+int motor_set_speed(const struct dfr_board *board, int motor, float speed);
 
 /*
  *
